user : update help command desc.
