+++ user/module/probe_openssl_lib.go
@@ -10,6 +10,38 @@ import (
 	"strings"
 )
 
+// initOpensslOffset initial BpfMap
+func (this *MOpenSSLProbe) initOpensslOffset() {
+	this.sslVersionBpfMap = map[string]string{
+
+		// openssl 1.1.1*
+		"OpenSSL 1.1.1a":     "openssl_1.1.1a_kern.o",
+		"OpenSSL 1.1.1b":     "openssl_1.1.1b-c_kern.o",
+		"OpenSSL 1.1.1c":     "openssl_1.1.1b-c_kern.o",
+		"OpenSSL 1.1.1d":     "openssl_1.1.1d-i_kern.o",
+		"OpenSSL 1.1.1e":     "openssl_1.1.1d-i_kern.o",
+		"OpenSSL 1.1.1f":     "openssl_1.1.1d-i_kern.o",
+		"OpenSSL 1.1.1g":     "openssl_1.1.1d-i_kern.o",
+		"OpenSSL 1.1.1h":     "openssl_1.1.1d-i_kern.o",
+		"OpenSSL 1.1.1i":     "openssl_1.1.1d-i_kern.o",
+		"OpenSSL 1.1.1j":     "openssl_1.1.1j-q_kern.o",
+		"OpenSSL 1.1.1k":     "openssl_1.1.1j-q_kern.o",
+		"OpenSSL 1.1.1l":     "openssl_1.1.1j-q_kern.o",
+		"OpenSSL 1.1.1m":     "openssl_1.1.1j-q_kern.o",
+		"OpenSSL 1.1.1n":     "openssl_1.1.1j-q_kern.o",
+		"OpenSSL 1.1.1o":     "openssl_1.1.1j-q_kern.o",
+		"OpenSSL 1.1.1p":     "openssl_1.1.1j-q_kern.o",
+		"OpenSSL 1.1.1q":     "openssl_1.1.1j-q_kern.o",
+		LinuxDefauleFilename: "openssl_1.1.1j-q_kern.o",
+
+		// openssl 3.0.*
+
+		// boringssl
+		"BoringSSL 1.1.1":      "boringssl_1.1.1_kern.o",
+		AndroidDefauleFilename: "boringssl_1.1.1_kern.o",
+	}
+}
+
 func (this *MOpenSSLProbe) detectOpenssl(soPath string) error {
 	f, err := os.OpenFile(soPath, os.O_RDONLY, os.ModePerm)
 	if err != nil {
